Clarification of keybinding
I just had a conversation with a co-worker and it was unclear that `=` was the key being assigned here and not being used as an assignment character. The initial thought was that `set-window-option synchronize-panes` was being assigned to `bind-key` and there was confusion about how an actual *key* was being assigned here.